requirements.txt
# Core dependencies
requests>=2.31.0
python-dotenv>=1.0.0
click>=8.1.0
flask>=2.0.0

# FastAPI Server dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
alembic>=1.12.0
pydantic>=2.4.0
pydantic-settings>=2.0.0
apscheduler>=3.10.0
pytz>=2023.3

# Development dependencies
pytest>=8.0.0
pytest-mock>=3.12.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
pytest-httpx>=0.30.0
httpx>=0.25.0
pyyaml>=6.0.0
